fix(dags): convert montant_dp to float type

+++ dags/cbcm/donnee_comptable/process.py
@@ -226,7 +226,9 @@ def process_demande_paiement(df: pd.DataFrame) -> pd.DataFrame:
     # Nettoyer les champs textuels
     txt_cols = ["centre_financier", "societe", "statut_piece", "montant_dp"]
     df = normalize_whitespace_columns(df, columns=txt_cols)
-    df["montant_dp"] = df["montant_dp"].replace({",": "."}).str.split().str.join("")
+    df["montant_dp"] = (
+        df["montant_dp"].replace({",": "."}).str.split().str.join("").astype(float)
+    )
 
     # Filtrer les lignes
     df = df.loc[df["statut_piece"] == "Comptabiliser"]
